requirements.txt
argon2-cffi==16.3.0
asn1crypto==0.24.0
bcrypt==3.1.4
bleach==2.1.2
boto==2.48.0
boto3==1.5.24
botocore==1.8.38
bz2file==0.98
cachetools==2.0.1
cairocffi==0.8.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
decorator==4.2.1
Django==2.0.2
django-debug-toolbar==1.9.1
djangorestframework==3.7.7
docutils==0.14
Faker==0.8.10
gensim==3.3.0
google-api-core==0.1.4
google-auth==1.4.1
google-auth-oauthlib==0.2.0
google-cloud-bigquery==0.30.0
google-cloud-core==0.28.0
google-cloud-storage==1.7.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
jellyfish==0.5.6
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
keyring==11.0.0
keyrings.alt==2.3
launchpadlib==1.10.5
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.4.0
numpy==1.14.0
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
pandas==0.22.0
pandas-gbq==0.3.1
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
plotly==2.3.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.3
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-debian==0.1.31
pytz==2018.3
pyxdg==0.26
PyYAML==3.12
pyzmq==16.0.4
recordlinkage==0.11.2
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
rest-pandas==1.0.0
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
SecretStorage==2.3.1
Send2Trash==1.4.2
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
smart-open==1.5.6
sqlparse==0.2.4
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
xcffib==0.5.1
zope.interface==4.4.3
